Refactor Teacher class to Staff class and rename modelTeacher to modelStaff only
