+++ src/main/java/seedu/address/logic/Logic.java
@@ -12,7 +12,7 @@ import seedu.address.model.modelCourse.Course;
 import seedu.address.model.modelFinance.Finance;
 import seedu.address.model.modelGeneric.ReadOnlyAddressBookGeneric;
 import seedu.address.model.modelStudent.Student;
-import seedu.address.model.modelTeacher.Teacher;
+import seedu.address.model.modelStaff.Teacher;
 import seedu.address.model.person.Person;
 import seedu.address.ui.SummaryPanel;
