+++ src/main/java/seedu/address/logic/commands/commandClear/ClearTeacherCommand.java
@@ -5,7 +5,7 @@ import static java.util.Objects.requireNonNull;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
-import seedu.address.model.modelTeacher.TeacherAddressBook;
+import seedu.address.model.modelStaff.TeacherAddressBook;
 
 /**
  * Clears the address book.
