+++ src/main/java/seedu/address/model/modelStaff/TeacherAddressBook.java
@@ -1,10 +1,7 @@
-package seedu.address.model.modelTeacher;
+package seedu.address.model.modelStaff;
 
 import static java.util.Objects.requireNonNull;
 
-import java.util.List;
-import javafx.collections.ObservableList;
-import seedu.address.model.modelCourse.Course;
 import seedu.address.model.modelGeneric.AddressBookGeneric;
 import seedu.address.model.modelGeneric.ReadOnlyAddressBookGeneric;
